# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "血癌照護管理平台",
        "description": "\n<h3>Application Interface Documentation for Blood Cancer Care Management Platform.</h3>\n<h4><a href='/'>Home Page</a></h4>\n                        ",
        "contact": {
            "name": "High Performance Distributed Systems Lab",
            "email": "nkust.ee.hpds@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/side-effects/account/{account}": {
            "get": {
                "tags": [
                    "副作用紀錄"
                ],
                "summary": "取得副作用紀錄",
                "description": "取得副作用紀錄",
                "operationId": "d9e7f638ee1c0a998c40ee9c2221bf45",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "個案帳號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "副作用紀錄編號",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "case_id": {
                                                    "description": "個案編號",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/effect"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/side-effects": {
            "post": {
                "tags": [
                    "副作用紀錄"
                ],
                "summary": "新增副作用紀錄",
                "description": "新增副作用紀錄",
                "operationId": "fd17324a04a937b9c88448afac5fc780",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "account"
                                        ],
                                        "properties": {
                                            "account": {
                                                "description": "個案帳號",
                                                "type": "string",
                                                "example": "user1"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/effect"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "副作用紀錄編號",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "case_id": {
                                                    "description": "個案編號",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/effect"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/account/{account}": {
            "get": {
                "tags": [
                    "每週任務"
                ],
                "summary": "取得每週任務",
                "description": "取得每週的任務相關資訊，如：週數、任務 id、該資料的獨立 id（在新增個案任務的完成狀態時會用到）",
                "operationId": "20996f7521f1fa148128a4ca35be0497",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "個案帳號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/tasks/state/{id}": {
            "patch": {
                "tags": [
                    "每週任務"
                ],
                "summary": "更新每週任務完成狀態",
                "description": "更新每週任務完成狀態。<p>先用 GET 該個案的所有任務後，再用任務 id 使用此方法。</p>",
                "operationId": "95b5c03bdbc7a20d6e5ca554383fa594",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "任務 id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "state"
                                ],
                                "properties": {
                                    "state": {
                                        "type": "string",
                                        "enum": [
                                            "completed",
                                            "uncompleted"
                                        ],
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/topics/{id}": {
            "get": {
                "tags": [
                    "任務題目"
                ],
                "summary": "取得任務題目",
                "description": "取得任務的題目相關資訊，如：題目 id。",
                "operationId": "d5f75a5ab8e665745c1ffc6281ed5609",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "任務 id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/topics/cases": {
            "post": {
                "tags": [
                    "任務題目"
                ],
                "summary": "新增個案任務題目完成狀態",
                "description": "新增個案每週任務題目完成狀態。<p>先 GET 任務 id（每週任務），再 GET 題目 id（任務題目），並以兩組 id 使用此方法設定，完成個案的答題狀態。</p>",
                "operationId": "2d2e284e9474eb54d2e1b0b7bb2466bd",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "case_task_id",
                                    "topic_id",
                                    "state"
                                ],
                                "properties": {
                                    "case_task_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "topic_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "state": {
                                        "type": "string",
                                        "enum": [
                                            "correct",
                                            "wrong"
                                        ],
                                        "example": "correct"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "effect": {
                "required": [
                    "date",
                    "symptom",
                    "severity",
                    "has_image"
                ],
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-3-6"
                    },
                    "symptom": {
                        "type": "string",
                        "example": "噁心"
                    },
                    "severity": {
                        "type": "integer",
                        "example": "5"
                    },
                    "has_image": {
                        "type": "integer",
                        "enum": [
                            "0",
                            "1"
                        ],
                        "example": "0"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary"
                    },
                    "caption": {
                        "type": "string",
                        "example": "Cation #1"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
    "security": [
        {
            "basicAuth": []
        }
    ],
...
}